+++ src/components/core/button.js
@@ -3,9 +3,9 @@ import styled, { css } from 'styled-components'
 export const Button = styled.button`
   display: flex;
   align-items: center;
-  background-color: #fff;
+  background-color: white;
+  color: var(--textSecondaryColor);
   border: 1px solid var(--boxBorderColor);
-  color: var(--textPrimaryColor);
   justify-content: center;
   padding: 0.5em 1em;
   text-align: center;
@@ -13,23 +13,35 @@ export const Button = styled.button`
   border-radius: 4px;
   font-size: 1rem;
   font-family: inherit;
-  ${props =>
-    props.disabled
-      ? css`
-          background-color: transparent;
-          > * {
-            opacity: 0.5;
-          }
-        `
-      : css`
-          cursor: pointer;
-          &:hover {
-            border-color: #b5b5b5;
-            color: #363636;
-          }
-        `}
+  &:hover {
+    cursor: pointer;
+    color: var(--textPrimaryColor);
+    border-color: var(--textMutedColor);
+  }
   &:active,
   &:focus {
     outline: 0;
+  }
+  ${props => props.primary && PrimaryButtonMixin}
+  ${props => props.disabled && DisabledButtonMixin}
+`
+
+const PrimaryButtonMixin = css`
+  background-color: var(--bestofjsOrange);
+  color: white;
+  border-color: var(--boxBorderColor);
+  &:hover {
+    background-color: var(--bestofjsPurple);
+    color: white;
+  }
+`
+
+const DisabledButtonMixin = css`
+  cursor: not-allowed;
+  background-color: transparent;
+  color: var(--textMutedColor);
+  &:hover {
+    color: var(--textMutedColor);
+    border-color: var(--boxBorderColor);
   }
 `