+++ src/root.tsx
@@ -10,23 +10,24 @@ import { fetchProjects } from './actions/entitiesActions'
 import createAuthApi from './api/auth/auth-api'
 import track from './helpers/track'
 import { fetchProjectsIfNeeded } from './actions/entitiesActions'
-import App from './App'
+import { App } from './app'
+import { useAppUpdateChecker } from 'app-update-checker'
+import { ToastContainer } from 'components/core/toast'
 
 export const Root = () => {
   const store = configureStore({})
   const authApi = createAuthApi({ dispatch: store.dispatch })
   const dependencies = { authApi }
 
-  useEffect(
-    () => {
-      store.dispatch(fetchProjects())
-    },
-    [store]
-  )
+  useEffect(() => {
+    store.dispatch(fetchProjects())
+  }, [store])
 
   return (
     <Router>
-      <AppWithRouter store={store} dependencies={dependencies} />
+      <ToastContainer>
+        <AppWithRouter store={store} dependencies={dependencies} />
+      </ToastContainer>
     </Router>
   )
 }
@@ -38,30 +39,29 @@ const AppWithRouter = props => {
   const location = useLocation()
   const history = useHistory()
 
+  useAppUpdateChecker({
+    interval: 5 * 60 * 1000, // check for updates every 5 minutes
+    isSimulationMode: false
+  })
+
   const {
     dependencies: { authApi },
     store
   } = props
 
-  useEffect(
-    () => {
-      authApi.start(history)
-    },
-    [authApi, history]
-  )
+  useEffect(() => {
+    authApi.start(history)
+  }, [authApi, history])
 
-  useEffect(
-    () => {
-      if (typeof window === 'undefined') return
-      window.scrollTo(0, 0)
-      track(location.pathname)
+  useEffect(() => {
+    if (typeof window === 'undefined') return
+    window.scrollTo(0, 0)
+    track(location.pathname)
 
-      // if the user is on the TOP page, reload data if data is stale
-      if (location.pathname === '/') {
-        store.dispatch(fetchProjectsIfNeeded())
-      }
-    },
-    [location, store]
-  )
+    // if the user is on the TOP page, reload data if data is stale
+    if (location.pathname === '/') {
+      store.dispatch(fetchProjectsIfNeeded())
+    }
+  }, [location.pathname]) // eslint-disable-line react-hooks/exhaustive-deps
   return <App {...props} />
 }